+++ CHANGELOG.md
@@ -4,7 +4,12 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
-[0.6.0] - 2025-02-09
+### Added
+
+- Card slots: `title`, `body` and `actions`
+- Card modifier with `side` and `image_full` properties
+
+## [0.6.0] - 2025-02-09
 
 ### Added